+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.15"
+__version__ = "0.12.16"
 
 
 def safe_commit() -> None:
